Start migration to lab-website-template (#1)
This PR copies over most content and information from the current (soon-to-be old) tislab.org website into the new version of the site based on lab-website-template. Some "TO DO" comments may not have been copied over. For now, I've only added Melissa's image and bio page. The other team members can be done in a future PR, possibly/hopefully by the team members themselves to reduce grunt-work on one person. Also, people's roles/titles (e.g. "Ontologist") aren't displayed right now. That is something I'm adding/changing in the template upstream.

This migration has caused me to rethink and refactor the template a lot (see the recent GitHub issues I posted on its repo). So hold tight while I make improvements to the template.

Going to merge this PR without review. I'm sure there will be lots of ideas and comments of things to change on the website, but ultimately that is something the lab will have to discuss, make issues for, and change on its own, and I don't want to hold up merging this because I want a basic preview to be available right now.