v1.1.2 (#178)
See https://github.com/greenelab/lab-website-template/discussions/177 .
v1.1.1 introduced a regression in `first-time-setup.yaml` when clearing
the `gh-pages` branch. I wanted to make the command simpler -- to simply
delete everything in the folder -- but I didn't think it through enough.
[There is a safeguard to prevent the command from
working](https://unix.stackexchange.com/questions/369530/rm-refusing-to-remove-or-directory-skipping),
and also even if it did work it would delete the `.git` folder which
would break things. Thus, this version reverts the command to how it was
before, manually specifying each `.` folder.